refactor: move api types to their own file
